+++ src/content/image_processor.cpp
@@ -306,7 +306,7 @@ static void detectImageDimensions(const std::vector<uint8_t>& blob, const std::s
     // JPEG: SOI + APP0 or SOF0
     if (blob[0] == 0xFF && blob[1] == 0xD8) {
         // Find SOF0/SOF2 marker
-        for (size_t i = 2; i < static_cast<int>(blob.size()) - 9; ++i) {
+        for (size_t i = 2; i < blob.size() - 9; ++i) {
             if ((blob[i] == 0xFF && (blob[i + 1] == 0xC0 || blob[i + 1] == 0xC2)) {
                 height = (blob[i + 5] << 8) | blob[i + 6];
                 width = (blob[i + 7] << 8) | blob[i + 8];
@@ -465,7 +465,7 @@ std::array<double, 1024> extractGrayscaleSamples(const std::vector<uint8_t>& blo
                         size_t off = pixel_offset
                             + static_cast<size_t>(actual_row) * static_cast<size_t>(row_stride)
                             + static_cast<size_t>(px) * 3;
-                        if (off + 2 <static_cast<int>(blob.size())) {
+                        if (off + 2 < blob.size()) {
                             uint8_t b = blob[off];
                             uint8_t g = blob[off + 1];
                             uint8_t r = blob[off + 2];
